requirements.txt
diffusers
matplotlib
pillow
torch
transformers